+++ cli/hc.test.ts
@@ -229,7 +229,7 @@ describe("main", () => {
 
     expect(status).toEqual(0);
     expect(out.at(-1).toString()).toContain(
-      '“{"total":12,"pass":12,"fail":0,"unimplemented":0}”',
+      '“{"total":16,"pass":16,"fail":0,"unimplemented":0}”',
     );
   });